fix storage error and related path

+++ app/Http/Controllers/AdminController.php
@@ -120,16 +120,19 @@ class AdminController extends Controller
                 // get admin data
                 $admin = auth()->user();
 
-                // Delete old file
-                $oldImagePath = 'upload/profiles_photos/' . $admin->image;
-                $oldThumbnailPath = 'upload/profiles_photos/thumbs/' . $admin->image;
+                // Delete the old image if it's not the default one
+                if ($admin->image !== 'profile.svg') {
 
-                if ($oldImagePath && Storage::disk('public')->exists($oldImagePath)) {
-                    Storage::disk('public')->delete($oldImagePath);
-                }
+                    $oldImagePath = 'upload/profiles_photos/' . $admin->image;
+                    $oldThumbnailPath = 'upload/profiles_photos/thumbs/' . $admin->image;
 
-                if ($oldThumbnailPath && Storage::disk('public')->exists($oldThumbnailPath)) {
-                    Storage::disk('public')->delete($oldThumbnailPath);
+                    if ($oldImagePath && Storage::disk('public')->exists($oldImagePath)) {
+                        Storage::disk('public')->delete($oldImagePath);
+                    }
+
+                    if ($oldThumbnailPath && Storage::disk('public')->exists($oldThumbnailPath)) {
+                        Storage::disk('public')->delete($oldThumbnailPath);
+                    }
                 }
 
                 // create image name
